0.4.1.3 Beta 10/14/2014 --User Changes  Size is no longer fixed to the window size.  Size scaling is done differently in the final render dialog. This fixes several bugs.  Remove Xml saving size from settings and options dialog, it no longer applies.  Final render can be broken into strips.  Set default save path to the desktop if none is found in the settings file.  Set default output size to 1920x1080 if none is found in the settings file.
--Bug Fixes
 Better memory size reporting in final render dialog.

--Code Changes
 Migrate to C++11, Qt 5.3.1, and Visual Studio 2013.
 Change most instances of unsigned int to size_t, and int to intmax_t.
 Add m_OrigPixPerUnit and m_ScaleType to Ember for scaling purposes.
 Replace some sprintf_s() calls in XmlToEmber with ostringstream.
 Move more non-templated members into RendererBase.
 Add CopyVec() overload that takes a per element function pointer.
 Add vector Memset().
 Replace '&' with '+' instead of "&amp;" in XmlToEmber for much faster parsing.
 Break strips rendering out into EmberCommon and call from EmberRender and Fractorium.
 Make AddAndWriteBuffer() just call WriteBuffer().
 Make AddAndWriteImage() delete the existing image first before replacing it.
 Add SetOutputTexture() to RendererCL to support making new textures in response to resize events.
 Remove multiple return statements in RendererCL, and replace with a bool that tracks results.
 Add ToDouble(), MakeEnd(), ToString() and Exists() wrappers in Fractorium.
 Add Size() wrapper in EmberFile.
 Make QString function arguments const QString&, and string with const string&.
 Make ShowCritical() wrapper for invoking a message box from another thread.
 Add combo box to TwoButtonWidget and rename.

+++ Source/Ember/CarToRas.h
@@ -39,7 +39,7 @@ public:
 	/// <param name="rasW">The width in pixels of the raster image/histogram</param>
 	/// <param name="rasH">The height in pixels of the raster image/histogram</param>
 	/// <param name="aspectRatio">The aspect ratio, generally 1</param>
-	CarToRas(T carLlX, T carLlY, T carUrX, T carUrY, unsigned int rasW, unsigned int rasH, T aspectRatio)
+	CarToRas(T carLlX, T carLlY, T carUrX, T carUrY, size_t rasW, size_t rasH, T aspectRatio)
 	{
 		Init(carLlX, carLlY, carUrX, carUrY, rasW, rasH, aspectRatio);
 	}
@@ -113,7 +113,7 @@ public:
 	/// <param name="rasW">The width in pixels of the raster image/histogram</param>
 	/// <param name="rasH">The height in pixels of the raster image/histogram</param>
 	/// <param name="aspectRatio">The aspect ratio, generally 1</param>
-	void Init(T carLlX, T carLlY, T carUrX, T carUrY, unsigned int rasW, unsigned int rasH, T aspectRatio)
+	void Init(T carLlX, T carLlY, T carUrX, T carUrY, size_t rasW, size_t rasH, T aspectRatio)
 	{
 		m_RasWidth = rasW;
 		m_RasHeight = rasH;
@@ -156,10 +156,10 @@ public:
 	/// <param name="cartY">The cartesian y</param>
 	/// <param name="rasX">The converted raster x</param>
 	/// <param name="rasY">The converted raster y</param>
-	inline void Convert(T cartX, T cartY, unsigned int& rasX, unsigned int& rasY)
+	inline void Convert(T cartX, T cartY, size_t& rasX, size_t& rasY)
 	{
-		rasX = (unsigned int)(m_PixPerImageUnitW * cartX - m_RasLlX);
-		rasY = (unsigned int)(m_RasLlY - (m_PixPerImageUnitH * cartY));
+		rasX = (size_t)(m_PixPerImageUnitW * cartX - m_RasLlX);
+		rasY = (size_t)(m_RasLlY - (m_PixPerImageUnitH * cartY));
 	}
 
 	/// <summary>
@@ -173,9 +173,9 @@ public:
 	/// <param name="cartX">The cartesian x</param>
 	/// <param name="cartY">The cartesian y</param>
 	/// <param name="singleBufferIndex">The converted single raster buffer index</param>
-	inline void Convert(T cartX, T cartY, unsigned int& singleBufferIndex)
+	inline void Convert(T cartX, T cartY, size_t& singleBufferIndex)
 	{
-		singleBufferIndex = (unsigned int)(m_PixPerImageUnitW * cartX - m_RasLlX) + (m_RasWidth * (unsigned int)(m_PixPerImageUnitH * cartY - m_RasLlY));
+		singleBufferIndex = (size_t)(m_PixPerImageUnitW * cartX - m_RasLlX) + (m_RasWidth * (size_t)(m_PixPerImageUnitH * cartY - m_RasLlY));
 	}
 
 	/// <summary>
@@ -190,9 +190,9 @@ public:
 	/// </summary>
 	/// <param name="point">The cartesian y</param>
 	/// <param name="singleBufferIndex">The converted single raster buffer index</param>
-	inline void Convert(Point<T>& point, unsigned int& singleBufferIndex)
+	inline void Convert(Point<T>& point, size_t& singleBufferIndex)
 	{
-		singleBufferIndex = (unsigned int)(m_PixPerImageUnitW * point.m_X - m_RasLlX) + (m_RasWidth * (unsigned int)(m_PixPerImageUnitH * point.m_Y - m_RasLlY));
+		singleBufferIndex = (size_t)(m_PixPerImageUnitW * point.m_X - m_RasLlX) + (m_RasWidth * (size_t)(m_PixPerImageUnitH * point.m_Y - m_RasLlY));
 	}
 
 	/// <summary>
@@ -221,8 +221,8 @@ public:
 	/// <summary>
 	/// Accessors.
 	/// </summary>
-	inline unsigned int RasWidth() const { return m_RasWidth; }
-	inline unsigned int RasHeight() const { return m_RasHeight; }
+	inline size_t RasWidth() const { return m_RasWidth; }
+	inline size_t RasHeight() const { return m_RasHeight; }
 	inline T OneRow() const { return m_OneRow; }
 	inline T OneCol() const { return m_OneCol; }
 	inline T PixPerImageUnitW() const { return m_PixPerImageUnitW; }
@@ -239,7 +239,7 @@ public:
 	inline T PadCarUrY() const { return m_PadCarUrY; }
 
 private:
-	unsigned int m_RasWidth, m_RasHeight;//The width and height of the raster image.
+	size_t m_RasWidth, m_RasHeight;//The width and height of the raster image.
 	T m_OneRow;//The distance that one raster row represents in the cartesian plane.
 	T m_OneCol;//The distance that one raster column represents in the cartesian plane.
 	T m_PixPerImageUnitW;//The number of columns in the raster plane that a horizontal distance of 1 in the cartesian plane represents. The higher the number, the more zoomed in.
